ai-model/spoilage_detection/train/requirements.txt
torch==2.1.0
torchvision==0.16.0
ultralytics==8.1.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
kaggle==1.5.16
matplotlib==3.8.2
seaborn==0.13.0



